Rename package refactor to send
Rename `life.qbic.business.notification.refactor` to `life.qbic.business.notification.send`
