+++ sample-notificator-app/src/main/groovy/life/qbic/samplenotificator/components/email/support/SupportEmailSender.java
@@ -10,7 +10,7 @@ import java.util.concurrent.TimeUnit;
 import java.util.stream.Collectors;
 import life.qbic.business.logging.Logger;
 import life.qbic.business.logging.Logging;
-import life.qbic.business.notification.refactor.FailureEmailSender;
+import life.qbic.business.notification.send.FailureEmailSender;
 import life.qbic.samplenotificator.components.email.EmailSendException;
 import life.qbic.samplenotificator.components.email.html.HtmlEmailSender;
