+++ sample-notificator-domain/src/main/groovy/life/qbic/business/notification/send/EmailSender.java
@@ -1,4 +1,4 @@
-package life.qbic.business.notification.refactor;
+package life.qbic.business.notification.send;
 
 import java.util.List;
 import java.util.function.Consumer;
