+++ sample-notificator-domain/src/main/groovy/life/qbic/business/notification/send/FailureEmail.java
@@ -1,4 +1,4 @@
-package life.qbic.business.notification.refactor;
+package life.qbic.business.notification.send;
 
 /**
  * <b>An email informing about system failure.</b>
